Autorelabel to trigger SELinux label repair
Image generation occurs on a system without SELinux
This seems to break SELinux labels in the filesystem

By having .autorelabel, if the VM is started with the distribution kernel
and SELinux is not disabled via kernel command line, SELinux labels will be
repaired. Thereafter SELinux can operate in Enforcing mode.

Signed-off-by: Hal Martin <hmartin@online.net>